MAJOR: Read file now returns pandas data
The good part: I know have a very insightful representation of my data.
The bad part: It seems to require a lot more storage, with the multiple
column levels (but not as much as the list approach I think)
So I traded speed for comfort.. not sure if it is worth it.